requirements.txt
python-telegram-bot
python-dotenv
dnspython
pymongo